Crate ethcore_blockchain

Blockchain database.

Modules

generator

Blockchain generator for tests.

Structs

BlockChain

Structure providing fast access to blockchain data.

BlockDetails

Familial details concerning a block

BlockNumberKey

Wrapper for block number used as a DB key.

BlockReceipts

Contains all block receipts.

CacheSize

Represents blockchain’s in-memory cache size in bytes.

Config

Blockchain configuration.

ExtrasInsert

Extra information in block insertion.

ImportRoute

Import route for newly inserted block.

TransactionAddress

Represents address of certain transaction within block

TreeRoute

Represents a tree route between from block and to block:

Traits

BlockChainDB

Database backing BlockChain.

BlockChainDBHandler

Generic database handler. This trait contains one function open. When called, it opens database with a predefined config.

BlockProvider

Interface for querying blocks by hash and by number.